identify and manage missing values
using the isnull() function, the missing values in the Sleep Disorder column are shown as it is the only column with missing values. It was decided that filling missing values with the string "None was the best option since it is realistic and neutral in this context instead of dropping rows.